refactor: rename pacchetto fzf/ → interface/
